+++ src/callbacks/postRender.ts
@@ -3,7 +3,6 @@ import { hasErrors } from "../classes/features/mandatory/misc/checkErrors/v";
 import { seededDeathPostRender } from "../features/mandatory/seededDeath/callbacks/postRender";
 import * as streakText from "../features/mandatory/streakText";
 import { fastTravelPostRender } from "../features/optional/major/fastTravel/callbacks/postRender";
-import { showDreamCatcherItemPostRender } from "../features/optional/quality/showDreamCatcherItem/callbacks/postRender";
 import { racePostRender } from "../features/race/callbacks/postRender";
 import * as customConsole from "../features/race/customConsole";
 import { mod } from "../mod";
@@ -25,9 +24,6 @@ function main() {
   // Major
   racePostRender();
   fastTravelPostRender();
-
-  // QoL
-  showDreamCatcherItemPostRender(); // 566
 
   // Other
   customConsole.postRender();