+++ src/callbacksCustom/postNewLevelReordered.ts
@@ -1,6 +1,5 @@
 import { game, log, ModCallbackCustom } from "isaacscript-common";
 import * as streakText from "../features/mandatory/streakText";
-import { showDreamCatcherItemPostNewLevel } from "../features/optional/quality/showDreamCatcherItem/callbacks/postNewLevel";
 import { racePostNewLevel } from "../features/race/callbacks/postNewLevel";
 import { mod } from "../mod";
 
@@ -24,7 +23,4 @@ function main() {
 
   // Major
   racePostNewLevel();
-
-  // QoL
-  showDreamCatcherItemPostNewLevel(); // 566
 }