+++ src/classes/features/optional/quality/showDreamCatcher/constants.ts
@@ -1,0 +1,1 @@
+export const DREAM_CATCHER_FEATURE_NAME = "ShowDreamCatcher";