+++ src/features/optional/major/fastTravel/checkStateComplete.ts
@@ -1,8 +1,8 @@
 import { EntityCollisionClass } from "isaac-typescript-definitions";
 import { game, getEffects, getPlayers } from "isaacscript-common";
+import { isDreamCatcherWarping } from "../../../../classes/features/optional/quality/showDreamCatcher/v";
 import { EffectVariantCustom } from "../../../../enums/EffectVariantCustom";
 import { FastTravelState } from "../../../../enums/FastTravelState";
-import { isDreamCatcherWarping } from "../../quality/showDreamCatcherItem/v";
 import { FADE_TO_BLACK_FRAMES, FRAMES_BEFORE_JUMP } from "./constants";
 import { setNewFastTravelState, setPlayersVisible } from "./setNewState";
 import { v } from "./v";